DEV-01 Refactoring and NgRX
Fixed the Time Travell Debugging for DevTools
