+++ package.json
@@ -26,11 +26,11 @@
     "@angular/platform-browser": "~11.0.5",
     "@angular/platform-browser-dynamic": "~11.0.5",
     "@angular/router": "~11.0.5",
-    "@ngrx/data": "^12.2.0",
+    "@ngrx/data": "^10.1.2",
     "@ngrx/effects": "^10.1.2",
     "@ngrx/entity": "^10.1.2",
-    "@ngrx/router-store": "^12.2.0",
-    "@ngrx/schematics": "^12.2.0",
+    "@ngrx/router-store": "^10.1.2",
+    "@ngrx/schematics": "^10.1.2",
     "@ngrx/store": "^10.1.2",
     "@ngrx/store-devtools": "^10.1.2",
     "bootstrap": "^4.3.1",
