+++ src/app/store/reducers/user/user.reducer.spec.ts
@@ -1,13 +1,13 @@
-import { reducer, initialState } from './user.reducer';
+// import { reducer, initialState } from './user.reducer';
 
-describe('User Reducer', () => {
-  describe('an unknown action', () => {
-    it('should return the previous state', () => {
-      const action = {} as any;
+// describe('User Reducer', () => {
+//   describe('an unknown action', () => {
+//     it('should return the previous state', () => {
+//       const action = {} as any;
 
-      const result = reducer(initialState, action);
+//       const result = reducer(initialState, action);
 
-      expect(result).toBe(initialState);
-    });
-  });
-});
+//       expect(result).toBe(initialState);
+//     });
+//   });
+// });
