+++ src/index.html
@@ -6,30 +6,30 @@
   <title>Khotso C Book Store</title>
   <base href="/">
   <meta name="viewport" content="width=device-width, initial-scale=1">
-    <!-- <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet"> -->
+     <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
   <style>
 
 /* @font-face {
   font-family: 'Material Icons';
   font-style: normal;
   font-weight: 400;
-  src: url(iconfont/MaterialIcons-Regular.eot); 
+  src: url(iconfont/MaterialIcons-Regular.eot);
   src: local('Material Icons'),
        local('MaterialIcons-Regular'),
        Old file: url(iconfont/MaterialIcons-Regular.woff2) format('woff2'),
-       load new file 
-       
-       url(iconfont/MaterialIcons-Regular.ttf) format('truetype'); 
+       load new file
+
+       url(iconfont/MaterialIcons-Regular.ttf) format('truetype');
 } */
 
   @font-face {
     font-family: 'Material Icons';
     font-style: normal;
     font-weight: 400;
-    src: local('Material Icons'), 
+    src: local('Material Icons'),
     local('MaterialIcons-Regular'),
-    url(assets/2fcrYFNaTjcS6g4U3t-Y5ZjZjT5FdEJ140U2DJYC3mY.woff2) format('woff2');
-  }  
+    /* url(assets/2fcrYFNaTjcS6g4U3t-Y5ZjZjT5FdEJ140U2DJYC3mY.woff2) format('woff2'); */
+  }
   .material-icons {
     font-family: 'Material Icons';
     font-weight: normal;
@@ -44,7 +44,7 @@
     direction: ltr;
     -moz-font-feature-settings: 'liga';
     -moz-osx-font-smoothing: grayscale;
-  } 
+  }
 </style>
   <link rel="icon" type="image/x-icon" href="favicon.ico">
