+++ tsconfig.json
@@ -27,6 +27,6 @@
     "strictInjectionParameters": true,
     "strictInputAccessModifiers": true,
     "strictTemplates": true,
-    
+
   }
 }
